Third time is the charm?
In any event, yet an attempt to exterminate those <IDLE> transactions that 
have been bugging us. 

Hopefully *this* patch will do what is expected (the extension is psycopg2-specific, 
since it is this module's transaction policy that is causing the problem. However
a similar approach may be warranted with a different db/backend as well. Adapt as 
needed).