Fix bug in module changelog 1.2 In changelog 1.2, change_handler_data had a constraint too much: PRIMARY KEY on evthdlr_key alone. This would disallow holes in the sequence, something that is actively used.
However, migrate_to_changelog_1_2 does not introduce this constraint,
so only databases created with changelog 1.2 is affected. This also
means that upgrading on these will give the user an error message.
The action DROP CONSTRAINT IF EXISTS would fix this, but this does not
work on some of the database servers (e.g. Postgres version < 9).