BofhCom: makes now use of the CerebrumCommunication interface, to easier swap the object which communicates with Cerebrum.
BofhForm: fixed the token to be active when the user is not logged on as well.
This requires the renewToken() to be called when logging on, as a good security
practice. Each form gets their unique token as well.

Changed the way of logging out users when the token has been changed. Instead of
a hardcoded call to the User object it now makes use of callbacks.

Changed some code to follow Pears code standard as well.