Code for synchronizing user, group and exchange data to AD. User sync and group sync tests seem to work just fine, but exchange sync hasn't been tested all that much. Thus, this code is not ready for production environment.
Note that this code is a full rewrite, and as such doesn't have much
in common with the AD-sync modules found in
Cerebrum/modules/(uio|hia|hiof|...), although it still depends on the
AD service command API in Cerebrum/modules/ADutilMixIn.py. The plan is
to make a better API, deprecate ADutilMixIn.py and rewrite the other
AD-syncs. Some beautiful day...