Changed email_remove_address so that users are able to delete some of their own addresses, i.e:
    - E-mail addresses that are not registered in one of the users affiliations'
      OU.

    - The default e-mail address domain, @ulrik.uio.no.

This mostly affects UiOs bofhd_auth.py, but a small change has been added to the
bofhd_uio_cmds.py (which other institutions inherit): we do not check that the
user are able to _create_ addresses in the same domain anymore, only that they
are able to delete in it. This will not affect other institutions as long as
only superusers are able to execute the command. If other users would be granted
access to the command, the auth for adding and deleting addresses have to differ
before this could create problems.