Changed the behaviour of accepting moderator invites through WebID. Now users see the invite for moderating a group and have to log on through Feide to accept the invitation. This both informs the user of what group the invite is about, and it means one less click for the user as they had to log on anyway. Previously you had to first log on, then click an accept button, without knowing what group you actually accepted to join.
The drawbacks of this change is that you can now search through invite keys by
brute force, and you can trick a user into automatically accepting a moderator
invite. The former would require a Feide account anyway to make use of the now
known invite id, and I don't think the latter is that dangerous, as it involves
no destructive actions, only a join.