Added first attempt of a client certificate check for CIS.
If cereconf.INDIVIDUATION_CLIENT_CERT is defined and either points to a client's
certificate file or a directory with all the clients' certificates, you wouldn't
be able to connect to the CIS without such a certificate. This hasn't been
tested enough, but quick tests show that it works as it should with the php
application (CICom).