The password service now checks that the account is active before the account is allowed to set a new password. Quarantines which do not block users from the service can be defined by:
    INDIVIDUATION_ACCEPTED_QUARANTINES = ('quarantine_autopassord', 
                                          'quarantine_svakt_passord',)

It also checks if the person/account is reserved by one of the groups defined in
in INDIVIDUATION_PASW_RESERVED. The superuser group is automatically added to
this list.

The list of available telephone types now contains the source_systems the
numbers should come from, as the types might change from system to system.
cereconf example:

    INDIVIDUATION_PHONE_TYPES = {
        'system_sap':   ('contact_mobile_phone', ),
        'system_fs':    ('contact_phone', 
                         'contact_mobile_phone',
                         'contact_private_mobile',)