Changing the behaviour of exceptions in Individuation.
Now, when errors occur, which should be given to the end user, a
CerebrumRPCExpcetion is raised. This exception only contains a code to a message
in a Messages object. The SoapIndividuationServer should translate this message
code into the correct message and language, and send it to the soap client
through a Fault object.