Makes an explicit db.close() after each soap call. Until now, the connectors haven't been closed after they're done. This could be the garbage collector's fault, since we are in threads.
It works ok in testing, the connections are closed, but the last connection
seems to always stay idle, according to 'select * from pg_stat_activity'. The
connection is closed and replaced by a new one when a new command is called. Why
it is like that, I don't know. 

This needs to be tested more before it can be considered stable. Especially when
dealing with parallell transactions.