Start importing mobile numbers from FS to the xml files. Import from xml to db will come.
Note that it also imports country and region codes for the numbers. This is to
be able to ignore numbers that is not from Norway when reading the xml files.
Also, if something exist in the country and/or region fields, FS do not validate
the numbers, which is an argument for ignoring such numbers in Cerebrum.