Added a new CIS webservice. This is for postmasters, giving they some feedback that bofhd can't be bothered with.
Note that the webservice has to be run with our own php client only, as it
doesn't contain any authentication. This is for now done through bofhd, which is
handled by our php client.

Note that the certificates hasn't been though much through yet, so it is not
ready for the production environment yet.