Added the php client for the new CIS webservice for postmasters.
The client makes use of bofhd for authentication, and requires that the user is
a member of either the postmaster, brukerreg or cerebrum group.

Note that no design has been added to the pages. Not sure if that is even
necessary, or if that comes in the way of output.