Added support for importing office addresses for people from SAP.
In test 905 persons did get an office address fully registered. 185 people did
not get their address populated, due to that either the building code or the
room number were blank. Note that the building codes have not been validated in
the import, that is a job for the eksport(s).