Created a new page for people to choose what their name should be set to. The requirements for the name is that the family name is equal to the already set primary family name, and the given names must all be found in the person's registered names.
To make this work in Brukerinfo, more data was needed from Cerebrum. The
"hidden" command person_name_suggestions(person_id) gives a person its name
varieties, together with what primary e-mail address this would create.

Also found a bugfix when testing this. First names on the form "F." would give
primary e-mail addresses on the form "f..lastname@domain", which is not valid.
This is now fixed in the wash of e-mail addresses.