Added a check for if cereconf.BOFHD_CHECK_DISK_SPREAD is set to anything, which might be on purpose if the instance is not using disks.
If BOFHD_CHECK_DISK_SPREAD is set, this change does nothing, but if it's blank,
this change avoids BofhdAuth to crash, and instead returns the same as if the
user's disk were not found. Depending on the instance' policy, this normally
makes users only modifiable by superusers - that is the auth method
'is_account_owner'.