Adjusted the SoapListener to accept the proper session IDs that is generated by twisted, and which the client has sent through its HTTP Cookie parameter. This is done by parsing the input data and put the session ID in as a last parameter.
A benefit is that only valid and existing sessions are accepted, a new session
ID is created if the old one is wrong. Another benefit is that we now don't have
to force the clients to send their session ID 

This has been tested through our PHP client (phplib/model/CICom.php), and works
both when bogus data and proper session IDs are sent. Other clients are not
tested yet.