Changed the behaviour for when groups marked for reservations gets deleted. It was previously considered an error, which blocked the service from being used. Now it only logs it as a warning and ignores the groups.
The drawback is that users that should be reserved from the service could become
unreserved due to a deleted group. They could still have the need of being
reserved, as they might have privileges outside of Cerebrum. The benefit of this
change is that the service does not block everyone from using the service if a
group is deleted.

Anyway, the service should not tag the groups in this way, but rather check for
a trait or anything similar instead.