Had to change how unknown exceptions were caught. Removed the call_wrapper and instead added the same functionality into different events that is triggered by rpclib.
This was needed to be able to let subclasses handle their own exceptions - which
the BasicSoapServer know nothing about - but still be able to return a generic
"Unknown Error" if an unknown exception was about to be returned to the client.

Exceptions were previously controlled in our call_wrapper:

    call_wrapper():
        try:
            return call_soap_method()
        except KnownExceptions:
            raise
        except Exception:
            raise UnknownException()

which disallowed subclasses to get hold of the exception. Now the route goes:

    call_wrapper():
        return call_soap_method()

    ...

    event_before_returning_an_exception(ctx):
        if ctc.out_error is not SoapFault:
            return UnknownErrorFault()

where the event function is called after any events added by the subclasses.
This is of course oversimplified. ;)