Fixed the call_wrapper to check for PermissionDenied as well. The use of events didn't work, as rpclib itself transforms a generic exception into a Fault, with str() of the exception. So that was a dead end, unfortunately.
The call_wrapper handles Cerebrum's Faults and exceptions, and now also
PermissionDenied, which it transforms into NotAuthorizedFault. To make this
work, this fault had to be moved from auth.py to the SoapListener. At least it
works.

Also did some cleanup in what to log, and what to call the event functions. Some
more cleaning is probably needed before the Group service could be considered
done, but now all the functionality is in place.