Bugfix. The pickle module does not like Cerebrum objects as input, as it then tries to put some weird exec data into the change_params, which fails at output. Normally it just stores the integer of a constant object, but not in one case.
The pickle does probably handle this in a smart way which I have no clue of, but
since the rest of the code just casts the objects to integers, I take the easy
way out and do it here as well.

This will avoid new bugs from cinfo_add and cinfo_del, and have checked the
change_params in Cerebrum/Entity.py, without finding any more such bugs, if it
is a bug.