Updated the forgotten password service at UiO with better feedback for users when they are reserved and should get the username sent to them by SMS. In addition the design were updated a bit after feedback from the web design team at USIT.
Added a new method in phplib to easier check that a text key exists, so that the
other instances will not be affected unless they define the required text.

Also updated the feedback messages on the server side, to give less explanations
to the client, as more of the explanation is now put at the client side.