Added support for importing mobile numbers' country and region codes. This is needed to be able to know if the number is for other countries than norway, and to later be able to support other countries, as the numbers then has to be concatenated.
Not sure if this is exactly what UiA wants, but it is easy to change the two
lines of code in import_FS to e.g. also support country codes of type +47.

Also, removed all the \t and cleaned up in the indentations in the import
script. Sorry, I just had to, to be able to read the code without getting too
confused.