Two bugfixes for giving superusers support for changing group owners.
The first is just adding a FormatSuggestion to the bofh command "group
change_owner", so that superusers are able to see the confirmation key that has
to be given to the new owner.

The other change is a bugfix, as the current group owner's entity_id is put in
the event's change_params. The problem is that it's not the group owner's id but
the authenticate operator's entity_id that is added. This is normally okay, but
not for the cases where a superuser is forcing a group owner change, where this
would raise an AssertionError when the new owner confirms the owner change.