Updated 'user create' and 'user promote_posix' to not ask for default file group, as UiO should only be using personal file groups from now on. The PosixUserUiOMixin is ignoring any given dfg anyway.
To avoid conflicts with other instances' bofhd commands, UiA has gotten a copy
of the original user_promote_posix. The different instances have their own
versions of the user_create command, but the prompt command is used by some.
They are however not affected, as they're not using (or specifying) "PosixUser"
in their bofhd commands, which makes the prompt method not making use of this
update.