To be able to make use of a generic group_create bofhd method which is different for UiO than the rest, I created a new class for common methods. This was to avoid blocking up UiO's command class with more complexity.
Note that this update requires that all bofhd classes that imports
bofhd_uio_cmds has to be updated. I've done this for UiA's class, just to see
how it could be done.

It's tested and works, but it should be reviewed and declared sane or insane, as
the use of the 'all_commands' dict might be a bit weird.