Migrated all the different bofhd_*_cmds classes to make use of the BofhdCommonMethods semisuperclass.
All instances should now:

- Inherit from BofhdCommonMethods if it's a regular instance. WebID should for
  instance not make use of this.
- In __init__, run super.__init__(), to fix superclass's all_commands setup.
- Copy in the superclass's all_commands, if not already set.
- Stop importing the first migrated command, group_create, from
  bofhd_uio_cmds.py. It's now automatically imported from BofhdCommonMethods.

This is only a small step towards a bofhd with a better hierarchy, but we now
have the BofhdCommonMethods class to put common instance methods, instead of
polluting bofhd_uio_cmds even more. Later on, we might want to make use for
Cerebrum.Utils.Factory to create the bofhd command classes, and e.g. have
different classes for group commands, person commands and user commands, but
that is a bit larger step. :)