Added a not that good bugfix for handling that some users could already be members of some groups, even though the cache says otherwize. The fix is to check the database for if the account is a member or not.
The downside of this commit, is that we now ask the database one more time for
group members, if we want to add a user to a group. It could probably be fixed
in a better way, but since we are working on new scripts for handling student
accounts I don't think I should dig much more in this. :)