Added support for adding persons to groups in AD. Persons that exist in groups with AD-spread, or in a subgroup of a group with AD-spread, will get its primary account added to the group in AD.
Note that if a person's primary account does not have the correct AD-spread, the
sync will find the person's most prioritized account that has the AD-spread and
use that account instead. This might be confusing for those few it happens for,
but the alternative would be to ignore the person from the group.