Changed the query for emnestudents to include those from the previous term until halfway into the current term. This is a workaround to make the students able to use Feide a while after they're done studying.
The change was checked against FSPROD.uio.no. The result was that another 1608
unique students got imported from FSPROD and into Cerebrum. It looks like most
of these was actually emnestudents last term, but this might not be the case for
all of them.

The change must be confirmed by STA, or jazz, before getting into production.

The actual request was for getting access in one year, this is only for up to
six months. We could of course change the query more, but I think we would
rather have the new studaut system in place instead of working more on this.