Changed the import FS for UiO to not remove old affiliations instantly, but are now waiting 365 days first, checking the date of when the affiliations was last updated by a script - the affiliations 'last_date'.
This is due to a request from STA, as they want all students to keep their
student affiliations for up to a year after ended studies, no matter if they
were active, drgrad, privatist, opptak or emnestud.

Note that EVU-students should not be affected by this, such affiliatons are
still removed at once.