Added back in UiO's design for Brukerinfo. It got somehow removed in rev 15602, when the design for HiNesna was created by me.
It looks like the cause of this error is that HiNesna's design directory was
copied from UiO's design dir, and I guess I forgot to NOT copy over the .ssh
directory as well.

The files are added back in as how they looked like from rev 15602. I didn't see
that any changes were added to them in later revisions. They didn't exist
anyway, so that would be even weirder...