Added a debug message to discover mismatches in various hashes for an account.
We discovered that some hashes of HA1 did not match an account's given password,
even though the other hashes for the account matched. This is weird and should
not happen, and so we should be notified if this happens any more.

It is not logged if the logger is not instantiated, though, as is the case if
only the core API is used. You would then not discover this problem.