Added back support for names by OUs.
Did spend all my sanity points just by finding the bug, and figuring out what it
was trying to do before, so I don't know if I've fixed the bug the way it should
be done. Someone else has to take a look at this.

The abcconf's use of "semi-constants" looks just weird to me.

<Name... /> -> 'ORG_NAME_NAME' -> 'OU_NAME_NAME' -> 'name' -> 'OU_NAME_NAME' -> ?

Previously this was hardcoded to making use of 'name' directly, which was
somewhat what I did here.