Bugfix in the OU API for TSD.
Updated the XML import script to support the XML format from Nettskjema. Minor
changes would come in the next few weeks, though. Nettskjema does not have
documentation about the XML format, so we have to make a few guesses when
parsing it. This is the initial attempt, just to have something to work with.