Merged the new AD-sync back into trunk.
Some files are removed from the old AD-sync, which has to be readded in a later
commit.

The sync is working, but contains bugs and misses some features, so this should
not be considered as being ready for production. It's more a cleanup before the
last pieces are hammered together.