CRB-110: Adding support for HomeMTA for ØFK, due to a problem with migrating to 2010.
Normally homeMTA is not used in Exchange 2010 and later, but the sync makes
the objects corrupted. Setting the homeMTA seems to fix the corruption, and
then ØFK is able to at least fix the object on their side.

This issue must be handled by us in a better way later, to avoid that ØFK has
to fix all new mailboxes.