Bugfixes in setting attributes. Discovered that attributes didn't get updated if they weren't already set in AD, by the current powershell command.
Also, included a more generic way of defining attributes, by defining their
content in the config, instead of having to hardcode the attribute names. This
would make it easier to add new or change attributes by sysadmins, instead of
having to develop each and every attribute. Only created for ContactInfo for
now.