Added much more functionality for setting AD attribute values through config.
By generating attributes somewhat generic, we don't have to develop for every
attribute that is requested by the instances, but sysadmin could instead set
them a bit more dynamically.

Removed most of UiA's local functionality for the AD sync, as the config now
handles most of it instead.

There are still some functionality missing, and a cleanup is strongly
recommended, but we're getting closer.