Extended the new AD sync with more config of attributes.
We now have most of the existing attributes as in UiA's current sync. Some bugs
remains, and more testing is needed.

The sysadmins are now able to set and modify more of the attributes, which will
hopefully give the instances faster changes in the sync, as long as the changes
are small.

Added a feature in the sync script for dumping out the Cerebrum data. This has
been requested by AD administrators to be able to see what data that should be
given, even before the AD environment is properly set up. This could then also
be used later, for debugging when upgrading.

Did some house cleaning while at it.