Added a new bofh command for affiliating different entities with a project.
Should for instance be used for connecting a host to a certain project when it
is not automatically created. This is rather a helper command for the
TSD-admins, for easier management. The command trait_set could be used e.g. for
groups, but you would then need to know that you must set the "target_id" to the
project's entity_id. Also, trait_set does not support hosts, and bofhd is too
messy to implement DnsOwners into the basic commands - don't want to take the
risk.