Changed the behaviour in the AD sync for how hosts are affiliated with projects.
Instead of syncing all hosts, we now only sync hosts that have a trait targeted
at a project OU. This is also a bit different in that hosts are now not
affiliated with projects through the subnet they are in.

We should by this be able to also separate projects by subnets in the future, as
several projects might need to share a vlan in the future.