Added a new attribute configuration for generic attributes that requires a callback with the entity itself.
This features might be needed in special cases in which we have no other
solution for and don't want to code in extra functionality. 

An example of usage could be Host entities to set DNS name:

    'DNSHostName': ConfigUtils.CallbackAttr(lambda ent: getattr(ent, 'hostname', '')),

Note that this requires that the 'hostname' attribute is set somewhere in the
code - this can not be fixed the config.