Moved the functionality for NisNetGroups into the regular group sync. The groups are now created and updated as regular groups, including the member lists.
We have a problem with the member lists, though, as we need to have the full
DistinguishedName to identify the members generically. Not all member type have
SAMAccountName, so we can't use it as an identifier. Need to fix this.

Cleaned up in the functionality while doing this. Now we are using the generic
commands Get-ADObject and New-ADObject instead of Get-ADGroup and New-ADGroup. A
few parameters have changed, but we get more generic code.