Changed how the ADclient is given the objectClass and makes use of it.
Previously, a map from entity_type to objectClass in AD was used. Now, this is
instead added freely as a string and could be specified per ADSync class, and
overruled by the configuration. We are now able to create objects of any classes
without having to care about the Cerebrum entity_types.