Started refactoring how group memberships are updated, for robustness.
Each member is instantiated by its own CerebrumEntity class, to be able to
calculate the Name and OU to be used for AD. Subclasses might even want to use
more data, depending on the structure used in AD.

As we in AD need to know the full DistinguishedName before we could add a
member, we need to ask AD about this since we don't always know the exact path.
This process could be optimised, as we could for instance take a first guess and
in most cases save time on that, or we could generate a full of objects before
starting the membership sync.

More work is needed, especially for cleaning up after the refactoring.