Updated the hostpolicy sync to include hosts as members.
Needed to fetch the DistinguishedName from AD, and convert the name a bit before
it could work.

This could, and should, be moved into the generic AD sync, but does not have
time for that until next year. Will be hardcoded for now, unfortunately. This
means that some of the config for this sync can not be changed without also
updating the hardcoded details.