Changed the AD sync into fetching data from AD in JSON format rather than CSV.
This makes us able to fetch more data types than only strings.

The python-json module were not implemented for python2.5, so had to include it
into Cerebrum/extlib/ for now, with a few small tweaks. Could be removed when
working in python2.6.

Had to raise the socket timeout quite a bit for this to work, as the JSON cmdlet
requires more from the AD server, so we need to have a bit more patience. At
least we get better data this way. :)