CRB-200: Added an attribute for setting person names.
Started on extending the configuration to support more than one value definition
per attribute, in case of entities which would need another value if the primary
value was not found. An example is nonpersonal accounts, which will not match a
PersonNameAttr, and would therefore need e.g. gecos.

Included a few bugfixes, too.