Extended how the attributes are modified when found in Cerebrum.
Now, the attribute list supports lists of attribute configurations, were the
first available is used. This was needed for DisplayName, which had to had to be
set in the priority:

1. The person's full name, if it's a personal account.
2. The gecos, if non-personal posix user.
3. Last restort, the username.

This would probably also be needed in other situations, so generalized it.

Some fetch-commands in ADSync.py needs to be updated before lists of configs
could be used.