Also included a check on sets when comparing attributes.
AD does not give us sets, but Cerebrum code might make use of this type. When
included through sorting, we are able to correctly compare them with AD and
avoid having to continously update such attributes.

The compare functionality might need to be updated, as we have discovered more
"features" in AD.