Included a class for running Update-Recipient, as in the old AD-sync.
Note that the functionality is not finished yet. The parameters for the cmdlet
are for instance not set, and the functionality has not been tested, except for
the syntax.

The Update-Recipient functionality, or its likewise in later Exchange versions,
should rather be processed through the EventHandler. In the old AD sync it will
not be rerun if it fails, but this is something the new EventHandler handles.