Changed adding group members to include indirect members.
Previously, only direct members were added for groups, and we would let any
parent group be included if it had the correct AD spread. Now, we include the
indirect members if the parent group does not have an AD spread, by flatten the
targeted group.

Since I'm bad at explaining, how about a figure? :)

Example on a situation in Cerebrum::

    +-------------------------+
    | GroupA (with AD-spread) |
    +-------------------------+
         |               |
    +---------+ +----------------------------+
    | MemberA | | GroupB (without AD-spread) |
    +---------+ +----------------------------+
                         |
                    +---------+
                    | MemberB |
                    +---------+

where GroupA, MemberA and MemberB is supposed to be in AD (have the correct AD
spreads), while GroupB should not be in AD.

In the previous AD sync, this would in AD become::

    +-------------------------+
    | GroupA (with AD-spread) |
    +-------------------------+
         |
    +---------+
    | MemberA |
    +---------+

While with this commit, the structure in AD would look like::

    +-------------------------+
    | GroupA (with AD-spread) |
    +-------------------------+
         |           |
    +---------+ +---------+
    | MemberA | | MemberB |
    +---------+ +---------+

I.e. the group is flattened in this situation. Note, however, that if GroupB had
the correct AD-spread, AD would have the same situation as in Cerebrum, which is
the same behaviour both in the old and the new AD sync.