Bugfix in how makedb.py ignores comments.
When makedb.py discovered long comments on single lines, the rest of the SQL
file were ignored. A few db-tables were therefore missing after setting up a new
Cerebrum environment.

CRB-388

We should change how db-tables and modules are related, but that have to happen
later. For now, we just need it to work.