Added initial tests for the AD2 sync.
Created a setup for being able to test more than this Cerebrum module, as we
will need it in the future. The motivation is to save work by having a common
setup which could be reused.

Note that not all the AD2 tests are working yet, and some are only skipped for
now, as they're not completed. This is only the first attempt at testing the AD2
sync.