Improves and stabilise the update of AD attributes.
Now, the AD attributes are still updated even if one of the attribute elements
are bogus. This will make AD refuse to update the rest of the attributes, so the
AD sync now works more aggressively by adding each element in a single command.
This is slow, but should not happen that often.

CRB-363