Fetch users from AD_LDAP instead of root DSN.
Giske's domain is also updated with other entities and not just Cerebrum. By
fetch all user objects, we are in danger of deactivating users that we shouldn't
touch.

Note: This has not been tested yet. I am not sure if we are able to specify OUs
in this way for the AD service.

CRB-297