Handle exceptions for when updating persons.
This avoids that the whole script crashes, e.g. if we get bogus data from FS.
Only the given person will not be updated in such scenarios. More info about
scenarios could be found in CRB-217.

Added a db.rollback in the exception handler to avoid db transaction blocks
which creates problems for the next person to be processed.