Add support for BofhdRequest for deactivations.
Now the script should work for UiO too, if --bofhdrequest is given. This would
make the script send delete requests for each account to BofhdRequest, which
process the account further, e.g. for archiving the home disk and imap data.

Added some bug fixes for input parameters and logging, too.