Fix: Enable certificate and hostname validation in the ad2-client
This change adds a https module that contains patched versions of
httplib.HTTPSConnection and urllib2.HTTPSHandler. It also introduces a
SSLContext-like object for configuring SSL parameters for the
https.HTTPSConnection object.

Also added ssl_match_hostname backport to extlib, to support hostname
verification.

This change requires the ssl module introduced in python 2.6. A backport of ssl
is needed for older versions of python.

CRB-468-ad2-mutual-authentication