Add mixins to make Cerebrum.Group.Group aware of other group types.
This adds mixin support for `Group.demote', `Group.get_extensions'. Added base
mixins for PosixGroup and DistributionGroup that can be added to the base group
to provide support for `Group.delete' when the group is really a PosixGroup or
DistributionGroup.

Note: This mixin will cause `DistributionGroup.delete()' and
`PosixGroup.delete()' to actually delete the Group, not just demote it.

posix-aware-group