Add support for quarantine change types in AD2 Quicksync
Quarantines are now handled for accounts in AD2 Quicksync. When a quarantine
change is detected, the account is locked or unlocked in AD, depending on the
current quarantine state in Cerebrum.

When multiple changes with the same (change type, subject id, dest id) is found,
only the latest will be executed (unless it fails -- then the previous will be
executed). Duplicate changes are marked as 'processed' in the changelog, unless
--change-ids is given.

Note that quarantine:del and quarantine:set are different change types. If an
account is quarantined, then un-quarantined, the account will be unlocked twice
in AD by quicksync.

Using quicksync with --change-ids will follow the same behaviour as regular
quicksync (i.e. only handle newest change in case of duplicates, ). Also, using
--change-id will no longer cause the sync to crash.
CRB-751-ad-quicksync-av-karantene-for-konto