Make AD-sync search for objects everywhere!!!
This is implemented as a fall-back. If an object is not found, we search
the entire tree. If we find more than one object with the same CN, we
abort and use the original CN.

This functionality can be decorated upon functions that take a
datastructure containing DistinguishedNames.

CRB-794: -ad-synk-uavhengig-av-plassering