AD-sync: Avoid erroneous group-membership removal.
When DistinguishedNames are removed from multivalued attributes (i.e.
group memberships), we check if the object exists somewhere else in the
catalogue. If there exists only one other CN that is identical to the CN
of the DN, we assume that this is the original object, and omit removal.

CRB-794: -ad-synk-uavhengig-av-plassering