Add date/fname to email in quarantine_accounts.py.
'${QUARANTINE_DATE}' in the template will now be substituted with the
date the quarantine will be set on the account.

'${FIRST_NAME}' in the template will now be substituted with the first
name (from system_cached) of the person owning the account.

If the patterns does not occur in the template, the date or the name
will not appear in the E-mail sent.

CRBD-213-automatikk-med-sperring-avvikling