Add trait for phone exchange, as they need it in AD.
Going over to using a trait instead of the AD attribute table, since the value
is meant to be used also in ABC, and is for generic use. Some of the extension
attributes were then not necessary.

More of the AD attribute types could also be removed, but that requires that we
have bofhd-functionality for HiØ to e.g. set titles and contact addresses
manually, which we don't have.

Part of CRBD-228