Moves change_type codes to core_tables
This moves the create/drop SQL design of change_type from mod_changelog.sql to
core_tables.sql.

Change type codes are used in the "core" classes in Cerebrum, so the codes must
exist, regardless of the design. Also, the codes will now be used for other
purposes than logging change-history to the database.

CRB-920-change-type-code