make target2psread_populate faster, from ~10 min to 18 seconds.
First of all lookup all accounts with exchange spread in one go
instead of looking them up while finding target_ids.

Also skip a useless extra EmailTarget clear+find.

A big benefit here is that the function now returns a set instead
of list, which is O(1) compared to lists O(n). As targ2spread is
searched in quite often, this makes a huge bonus for other functions,
in addition to the almost 10 minutes saved for this function alone.