Adds a mixin class for PasswordHistory.
The mixin class should make it easier to add PasswordHistory to a given class.
It also removes dependency from Cerberum.Account to PasswordHistory.

Also adds an API method to Cerberum.Account – password_good_enough - that should
be used to check if a password is fit for use with an Account. The
PasswordHistoryMixin use this API method to perform the check against old
passwords.

CRB-192-password-check-cleanup