Add support for specifying a DC server to talk with
This overrides what Domain Controller we should talk with, and should only be
used in special situations. Currently, this has until now only happened with
HiH, which have two domains we should sync with, and only one WinRM server. We
need to override the DC to be able to manage the domain which the WinRM server
is not located in.

CRB-972 and CRB-976