Factory: split assemble and cereconf
New method Factor.make_class() will assemble a class given a list of
classes. Factory.get() now calls make_class to do its thing. This
will make localized factories easier.

The big question is, however, which direction Factory is taking. As of
today, Factor.get() is growing every time we need to add a new CLASS_X
to cereconf, but configuration should be telling the system what
features you want, not a list of classes.

The motivation behind this change is that cereconf is not versatile
enough. You want to communicate with some other end point, and different
software versions/configurations of this end point must be negotiated.
One solution might be to have factory negotiate, and return a proper
class for the job.