Factory.get("FS") → C.m.no.access_FS.make_fs()
Important: This new function will disregard cereconf.CLASS_FS, and use
cereconf.FS_MODULE in stead. Nobody seems to use mixins for FS, so this
should not be too much trouble.

The new implementation asks FS for version number, and finds the best
match in the configured module (defaults to access_FS).