UiA/process SAP affs: Remove deleted persons
If a person is no longer exported, populate_affiliation() is no longer
called, and write_db will not delete affiliations. Fix this by calling
delete_affiliation ourselves.

Note: Do we need to log removal of sap aff? If it is enough with db
entries, we could possibly make this code more elegant.