Ignore accounts without AD spread when processing quarantines from CL
This is a quickfix for CRB-1368, but only for quarantines and not for passwords.
Ignoring some quarantines are safe to do, since the fullsync cleans up in edge
cases, while passwords are only handled by the quicksync (except when creating
accounts).

This will hopefully speed up the quicksync when there a lot of quarantine events
for accounts without AD spread. We've discovered that it's a lot of them in
prod...