Enforce encoding on host parameter for urllib2.
Sometimes you kinda pass in a unicode(hostname) to a urllib2.Request. You
also pass a lot of data that is encoded as UTF-8.

Urllib/httplib/shitlib constructs a HTTP message, concatenated by a
Unicode object and a regular string. The result is a Unicode object.
Good fucking luck encoding that to ASCII.

DIRTY DIRTY URLLIB BURN BURN BURN!