Add support for group type/category, i.e. distribution
Seems like we haven't been able to create regular distribution groups, only
security groups. This has not been needed until now, as instances have used
security groups for mail as well.

This supports creating pure distribution groups, but not sure if we get errors
later in the sync, e.g. in functionality regarding to SIDs, which doesn't exist
for distgroups.

For NIH in CRB-1591.